src/requirements.txt
autocfg==0.0.8
certifi==2023.7.22
charset-normalizer==3.2.0
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.42.1
gluoncv==0.10.5.post0
graphviz==0.8.4
humanfriendly==10.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.7.2
mpmath==1.3.0
mxnet==1.9.1
nudenet==3.0.1
numpy==1.25.2
onnxruntime==1.15.1
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
portalocker==2.7.0
protobuf==4.24.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.2
six==1.16.0
sympy==1.12
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
yacs==0.1.8
